#131 - remove unused classes in hyperledger, move classes folder to ethereum
