--- src/main/java/blf/blockchains/hyperledger/reader/HyperledgerBlock.java
@@ -1,8 +1,0 @@
-package blf.blockchains.hyperledger.reader;
-
-/**
- * HyperledgerBlock
- */
-public abstract class HyperledgerBlock implements Iterable<HyperledgerTransaction> {
-
-}
