--- src/main/java/blf/blockchains/hyperledger/reader/HyperledgerClient.java
@@ -1,8 +1,0 @@
-package blf.blockchains.hyperledger.reader;
-
-/**
- * HyperledgerClient
- */
-public interface HyperledgerClient {
-
-}
