--- src/main/java/blf/blockchains/hyperledger/reader/HyperledgerLogEntry.java
@@ -1,8 +1,0 @@
-package blf.blockchains.hyperledger.reader;
-
-/**
- * EthereumLog
- */
-public abstract class HyperledgerLogEntry {
-
-}
