--- src/main/java/blf/blockchains/hyperledger/reader/HyperledgerTransaction.java
@@ -1,8 +1,0 @@
-package blf.blockchains.hyperledger.reader;
-
-/**
- * EthereumTransaction
- */
-public abstract class HyperledgerTransaction {
-
-}
