+++ src/main/java/blf/configuration/LogEntrySignatureSpecification.java
@@ -4,7 +4,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import blf.blockchains.hyperledger.classes.EthereumLogEntrySignature;
+import blf.blockchains.ethereum.classes.EthereumLogEntrySignature;
 import io.reactivex.annotations.NonNull;
 
 /**
